+++ lua/src/loslib.c
@@ -154,7 +154,7 @@ static int os_date (lua_State *L) {
       if (*s != '%' || *(s + 1) == '\0')  /* no conversion specifier? */
         luaL_addchar(&b, *s);
       else {
-        size_t reslen;
+        unsigned int reslen;
         char buff[200];  /* should be big enough for any conversion result */
         cc[1] = *(++s);
         reslen = strftime(buff, sizeof(buff), cc, stm);
